#10 Implement readParticipant hook, frontend for competition participation, reimplement addCompetition * Turn new-competition into a TypeScript file

+++ convex/participant.ts
@@ -1,5 +1,5 @@
-import { mutation } from './_generated/server'
-import { v } from 'convex/values'
+import { mutation, query } from './_generated/server'
+import { ConvexError, v } from 'convex/values'
 import { verifyUser } from './user'
 import {
   findTeamOfUser,
@@ -7,6 +7,14 @@ import {
   validateTeamJoinRequest,
 } from './lib/team'
 import { RequestValidity } from '../shared/info'
+
+export const readParticipant = query({
+  args: { competitionId: v.id('competitions') },
+  handler: async ({ db, auth }, { competitionId }) => {
+    const user = await verifyUser(db, auth)
+    return await findTeamOfUser(db, user, competitionId)
+  },
+})
 
 /**
  * Create a new empty team for a competition and adds the user to the competition through that team
@@ -19,14 +27,21 @@ export const joinCompetition = mutation({
 
     const competition = await db.get(id)
     if (!competition) {
-      throw new Error('That competition does not exist')
+      throw new ConvexError('That competition does not exist')
     }
 
     // Stop users who have left the competition and banned users from rejoining the competition
     if (competition.banned.includes(user._id)) {
-      throw new Error('You are not permitted to enter this competition')
+      throw new ConvexError('You are not permitted to enter this competition')
     }
 
+    // Check if the participant is already in the competition
+    const userTeam = await findTeamOfUser(db, user, competition._id)
+    if (userTeam) {
+      throw new ConvexError('You already entered this competition')
+    }
+
+    // Add user to competition and create new team
     const team = await db.insert('teams', {
       competition: id,
       joinRequests: [],
@@ -52,14 +67,14 @@ export const requestJoin = mutation({
     const user = await verifyUser(db, auth)
     const inviterTeam = await db.get(id)
     if (!inviterTeam) {
-      throw new Error('The inviting team does not exist')
+      throw new ConvexError('The inviting team does not exist')
     }
 
     const requestValidity = await validateTeamJoinRequest(db, inviterTeam, user)
     switch (requestValidity) {
       // Repeat request
       case RequestValidity.REQUESTED:
-        throw new Error('The request has already been made')
+        throw new ConvexError('The request has already been made')
       case RequestValidity.VALID:
         // Record the join request
         inviterTeam.joinRequests.push({
@@ -79,7 +94,8 @@ export const requestJoin = mutation({
 /**
  * Records a join request with the team's consent to an invitee OR adds a user to the team
  *
- * Postcondition: If a join request by the joiner is already made, adds the user to the team because both sides have consented.
+ * Postcondition: If a join request by the joiner is already made, adds the user to the team because both sides have
+ * consented.
  * If no such join request exists, records the join request.
  *
  * @param joinerId The user who is not on the team
@@ -90,17 +106,20 @@ export const inviteToTeam = mutation({
   handler: async ({ db, auth }, { joinerId, competitionId }) => {
     const inviter = await verifyUser(db, auth)
     const inviterTeam = await findTeamOfUser(db, inviter, competitionId)
+    if (!inviterTeam) {
+      throw new ConvexError('The user is not in that competition')
+    }
 
     const joiner = await db.get(joinerId)
     if (!joiner) {
-      throw new Error('The user getting invited does not exist')
+      throw new ConvexError('The user getting invited does not exist')
     }
 
     // Add user to team if the join request was in
     const validation = await validateTeamJoinRequest(db, inviterTeam, joiner)
     switch (validation) {
       case RequestValidity.INVITED:
-        throw new Error('The invite was already made')
+        throw new ConvexError('The invite was already made')
       case RequestValidity.VALID:
         // Record the join request
         inviterTeam.joinRequests.push({
@@ -118,7 +137,12 @@ export const inviteToTeam = mutation({
 })
 
 /**
- * Leaves a competition
+ * Permanently removes a user from a competition
+ *
+ * Preconditions:
+ * - User is in the competition
+ * - The competition exists
+ *
  * Postconditions:
  * - User is banned from competition IFF the team was multi-person
  * - User is unassigned from team he was last on
@@ -127,17 +151,21 @@ export const inviteToTeam = mutation({
  * - ALl join requests for team no longer exist IFF the team is a singleton
  * - The team the user was on no longer exists IFF the team is a singleton
  * @param id The id of the competition
+ * @throws ConvexError If not all preconditions are met
  */
 export const leaveCompetition = mutation({
   args: { id: v.id('competitions') },
   handler: async ({ db, auth }, { id }) => {
     const competition = await db.get(id)
     if (!competition) {
-      throw new Error('That competition does not exist')
+      throw new ConvexError('That competition does not exist')
     }
 
     const user = await verifyUser(db, auth)
     const team = await findTeamOfUser(db, user, id)
+    if (!team) {
+      throw new ConvexError('User is not in that competition')
+    }
 
     // Remove the user from the team and the competition
     await db.delete(team.userMembership)
